Update script.js
The whole code revolves around generate, so I made generate as readable as possible. All of the things generate does are done in separate functions to make generate as clean as possible.

Named variables and functions with more readable names. I used  things like "passwordLength" for the password length instead of confirmLength, etc.